requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.3
colorama==0.4.4
common==0.1.2
dataclasses==0.8
ddt==1.4.4
decorator==5.1.0
et-xmlfile==1.0.1
Flask==2.0.2
HTMLTestRunner-Python3==0.8.0
idna==2.10
importlib-metadata==3.7.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
jsonpath==0.82
keyboard==0.13.5
MarkupSafe==2.0.1
openpyxl==3.0.7
packaging==20.9
parameterized==0.8.1
pluggy==0.13.1
py==1.10.0
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.2
PyYAML==5.4.1
requests==2.25.1
scripts==2.0
selenium==3.141.0
self==2020.12.3
six==1.16.0
stua==0.2
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==2.0.2
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zipp==3.4.1
